securedrop-admin: do not hide stack trace when -v
When an error happens it is good to have a terse message and not
confuse the casual admin with a stacktrace. However it hides the real
cause of the error when debugging. For instance:

$ ./securedrop-admin sdconfig
ERROR: Pre-requisites not in place, re-run command with 'setup'

If in verbose mode, also display the stack trace to show the root
cause of the problem:

$ ./securedrop-admin -v sdconfig
ERROR: Pre-requisites not in place, re-run command with 'setup'
Traceback (most recent call last):
  File "./securedrop-admin", line 341, in <module>
    args.func(args)
  File "./securedrop-admin", line 49, in sdconfig
    activate_venv(args)
  File "./securedrop-admin", line 41, in activate_venv
    execfile(VENV_ACTIVATION, dict(__file__=VENV_ACTIVATION))
IOError: [Errno 2] No such file or directory: '/securedrop/.venv/bin/activate_this.py'